requirement.txt
acstools==3.2.0
alabaster==0.7.12
asdf==2.6.0
asteval==0.9.18
astroimtools==0.2
astropy==4.0.1.post1
astroquery==0.4
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
batman-package==2.4.6
beautifulsoup4==4.9.1
bleach==3.1.5
blessed==1.17.6
bokeh==2.1.0
brotlipy==0.7.0
calcos==3.3.9
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.4.1
corner==2.1.0
costools==1.2.3
crds==7.5.0.0
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.20
cytoolz==0.10.1
d2to1==0.2.12
dask @ file:///tmp/build/80754af9/dask-core_1592842333140/work
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
drizzlepac==3.1.6
-e git+https://github.com/dfm/emcee.git@91bf7c23739c5765c25198f5f7b387391556c6c1#egg=emcee
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
fitsblender==0.3.6
future==0.18.2
gatspy==0.3
ginga==3.0.1
gwcs==0.13.0
h5py==2.10.0
htc-utils==0.1.0
html5lib==1.0.1
hypothesis==5.16.1
idna==2.9
imageio==2.8.0
imagesize==1.2.0
imexam==0.9.1
importlib-metadata==1.6.1
iniconfig==1.0.1
inquirer==2.7.0
ipykernel==5.3.0
ipython @ file:///tmp/build/80754af9/ipython_1592348391367/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi @ file:///tmp/build/80754af9/jedi_1592841914522/work
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jwxml==0.3.0
keyring==21.1.1
kiwisolver==1.2.0
lmfit==1.0.1
lockfile==0.12.2
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592406092619/work
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.4.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nictools==1.1.5
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
numpydoc==1.0.0
olefile==0.46
openpyxl==3.0.3
optimparallel==0.1.1
packaging==20.4
pandas @ file:///tmp/build/80754af9/pandas_1592841660978/work
pandocfilters==1.4.2
Parsley==1.3
parso==0.7.0
patsy==0.5.1
pep517==0.8.2
pexpect==4.8.0
photutils==0.7.2
pick==1.0.0
pickleshare==0.7.5
Pillow==7.1.2
pkgconfig==1.4.0
pluggy==0.13.1
poppy==0.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.2
pycparser==2.20
pyds9==1.9.dev211
pyFFTW==0.9.2
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pysiaf==0.9.0
PySocks==1.7.1
pysynphot==0.9.14
pytest==6.0.1
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.7.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
readchar==2.0.1
reftools==2.0.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
scikit-image==0.16.2
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1592502882723/work
scipy==1.4.1
SecretStorage==3.1.2
semantic-version==2.8.4
Send2Trash==1.5.0
setuptools-scm==4.1.2
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
specutils==1.0
spherical-geometry==1.2.18
Sphinx @ file:///tmp/build/80754af9/sphinx_1592842202926/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
statsmodels==0.12.2
stginga==1.1
stistools==1.3.0
stregion==1.1.6
stsci.convolve==2.2.2
stsci.distutils==0.3.8.dev0
stsci.image==2.3.3
stsci.imagemanip==1.1.4
stsci.imagestats==1.6.2
stsci.ndimage==0.10.3
stsci.skypac==1.0.5
stsci.stimage==0.2.4
stsci.tools==3.6.0
stwcs==1.5.3
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.50.2
traitlets==4.3.3
tser-tools==0.1.dev0
tshirt @ file:///home/kglidic/Software/tshirt
tweakwcs==0.6.4
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1592847887441/work
uncertainties==3.1.4
urllib3==1.25.9
wcwidth==0.2.4
webbpsf==0.9.0
webencodings==0.5.1
wfc3tools==1.3.5
wfpc2tools==1.0.4
widgetsnbextension==3.5.1
zipp==3.1.0
